added orders and stripe api's

+++ src/main/java/co/turing/dto/response/CartTotalAmount.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class CartTotalAmount {
-    int totalAmount;
+    Double totalAmount;
 }
